launchTest boolean issue fixed

+++ app/src/main/java/com/bluetriangle/bluetriangledemo/DemoApplication.kt
@@ -1,6 +1,7 @@
 package com.bluetriangle.bluetriangledemo
 
 import android.app.Application
+import android.util.Log
 import com.bluetriangle.analytics.BlueTriangleConfiguration
 import com.bluetriangle.analytics.Tracker
 import com.bluetriangle.android.demo.tests.HeavyLoopTest
@@ -12,8 +13,10 @@ class DemoApplication : Application() {
         lateinit var tinyDB: TinyDB
 
         fun checkAndRunLaunchScenario(scenario:Int) {
+            val launchTest = tinyDB.getBoolean(KEY_LAUNCH_TEST)
             val launchTestScenario = tinyDB.getInt(KEY_LAUNCH_SCENARIO, 1)
-            if(launchTestScenario == scenario) {
+            Log.d("DemoApplication", "checkAndRunLaunchScenario: isLaunchTest: $launchTest, Scenario: $launchTestScenario")
+            if(launchTest && launchTestScenario == scenario) {
                 tinyDB.remove(KEY_LAUNCH_TEST)
                 tinyDB.remove(KEY_LAUNCH_SCENARIO)
                 HeavyLoopTest(3L).run()
@@ -29,6 +32,7 @@ class DemoApplication : Application() {
 //        if (!siteId.isNullOrBlank())
 //            intTracker(siteId) //bluetriangledemo500z
         val launchTest = tinyDB.getBoolean(KEY_LAUNCH_TEST)
+        Log.d("DemoApplication", "Application.onCreate: Launch Test: $launchTest")
         if(launchTest) {
             val siteId = tinyDB.getString(KEY_SITE_ID)
             val anrDetection = tinyDB.getBoolean(KEY_ANR_ENABLED)
